moved CycleList to utils.collection
